requirements.txt
Flask>=2.3,<3.0
Flask-SQLAlchemy>=3.1,<4.0
Flask-Migrate>=4.0,<5.0
Flask-Cors>=4.0,<5.0
Flask-Limiter>=3.8,<4.0

# Kimlik ve güvenlik
Flask-JWT-Extended>=4.6,<5.0
PyJWT>=2.9,<3.0
python-dotenv>=1.0,<2.0
argon2-cffi>=23.1,<24.0
passlib>=1.7,<2.0
cryptography>=44,<45
pyotp>=2.9,<3.0
pydantic>=2.8,<3.0
loguru>=0.7,<0.8

# Görev kuyrukları / cache
Celery>=5.3,<6.0
redis>=5.0,<6.0

# HTTP istemcileri
requests>=2.32,<3.0
httpx>=0.28,<1.0

# Metrikler (opsiyonel / /metrics uç noktası için)
prometheus-client>=0.16,<1.0

# Realtime (kullanıyorsan bırak; kullanmıyorsan ikisini kaldır)
Flask-SocketIO>=5.3,<6.0
eventlet>=0.33,<0.34

# Admin SPA dependencies
itsdangerous==2.2.0      # CSRF token üretimi (admin)
marshmallow==3.21.1      # Şema/validasyon (opsiyonel)
